Update test to be compatible with UIAutomator2
With the UIAutomator2, elements are less robust than they are with
UIAutomator. When the view changes, it can be necessary to re-find
elements. That is why we locate the resultField element immediately
before checking for it's contents. This change updates BasicTestSetup
to work with the UIAutomator2 backend.

Note that web testing does *not* currently work with UIAutomator2 backend.